refactoring src packages
