+++ src/domain/Convenente.ts
